Abstract logic for parsing PDF and getting definitions for words out of getRareWords
