+++ src/library/helpers/getRareWords.ts
@@ -1,28 +1,12 @@
-import { FileSystemUploadType, uploadAsync } from "expo-file-system";
+import type { CorpusWord } from "../../types/dataTypes";
 
-import type { DefinitionWord } from "../../types/dataTypes";
-
-const getRareWords = async (
-  fileData: string | File
-): Promise<DefinitionWord[]> => {
-  const URL = "https://readvocab-backend-production.up.railway.app/words";
-  if (fileData instanceof File) {
-    const formData = new FormData();
-    formData.append("pdf", fileData);
-    const response = await fetch(URL, {
-      method: "POST",
-      body: formData,
-    });
-    const data = await response.json();
-    return data;
-  } else {
-    const response = await uploadAsync(URL, fileData, {
-      fieldName: "pdf",
-      httpMethod: "POST",
-      uploadType: FileSystemUploadType.MULTIPART,
-    });
-    const data = JSON.parse(response.body);
-    return data;
-  }
+const getRareWords = async (words: string[]): Promise<CorpusWord[]> => {
+  const URL = "http://0.0.0.0:3000/rareWords";
+  const response = await fetch(URL, {
+    method: "POST",
+    body: JSON.stringify(words),
+  });
+  const data = await response.json();
+  return data;
 };
 export default getRareWords;
